+++ libvectors/srcs/get_vector.c
@@ -6,13 +6,13 @@
 /*   By: rotrojan <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/02/03 17:24:52 by rotrojan          #+#    #+#             */
-/*   Updated: 2020/02/03 17:27:14 by rotrojan         ###   ########.fr       */
+/*   Updated: 2020/02/16 23:09:20 by rotrojan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "vectors.h"
 
-t_vector	get_vector(float x, float y, float z)
+t_vector	get_vector(double x, double y, double z)
 {
 	t_vector vec;
